move c2usb interfacing code under the right sources
