+++ device/src/pin_wiring.c
@@ -6,7 +6,6 @@
 #include "keyboard/leds.h"
 #include "keyboard/oled/oled.h"
 #include "logger.h"
-#include "usb/usb.h"
 #include <zephyr/drivers/adc.h>
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/shell/shell.h>
@@ -17,7 +16,6 @@
 #include "host_connection.h"
 #include "thread_stats.h"
 #include "trace.h"
-#include "usb_compatibility.h"
 #include "mouse_keys.h"
 #include "config_manager.h"
 #include <zephyr/shell/shell_backend.h>
