+++ right/src/hid/CMakeLists.txt
@@ -1,0 +1,10 @@
+target_sources(${PROJECT_NAME} PRIVATE
+    app_base.cpp
+    command_app.cpp
+    controls_app.cpp
+    gamepad_app.cpp
+    keyboard_app.cpp
+    mouse_app.cpp
+    usb_compatibility.cpp
+    usb.cpp
+)
